Use `int16` in doctests for consistent reprs
Windows defaults to `int32` so won't display those `dtype`s. Whereas
UNIX defaults to `int64` and so won't display those `dtype`s. This
creates mismatches between the representations in docstrings between the
two platforms. For simplicity just use `int16`, which both platforms
will represent the same. This should ensure consistency with doctests.

+++ numcodecs/astype.py
@@ -28,10 +28,10 @@ class AsType(Codec):
     >>> x = np.arange(100, 120, 2, dtype=np.int8)
     >>> x
     array([100, 102, 104, 106, 108, 110, 112, 114, 116, 118], dtype=int8)
-    >>> f = numcodecs.AsType(encode_dtype=x.dtype, decode_dtype=np.int64)
+    >>> f = numcodecs.AsType(encode_dtype=x.dtype, decode_dtype=np.int16)
     >>> y = f.decode(x)
     >>> y
-    array([100, 102, 104, 106, 108, 110, 112, 114, 116, 118])
+    array([100, 102, 104, 106, 108, 110, 112, 114, 116, 118], dtype=int16)
     >>> z = f.encode(y)
     >>> z
     array([100, 102, 104, 106, 108, 110, 112, 114, 116, 118], dtype=int8)
